resolve sonnar cloud code smell
