+++ frontend/foodbook/src/containers/RealCalendar/RealCalendar.js
@@ -5,35 +5,12 @@ import { withRouter } from 'react-router';
 import 'semantic-ui-css/semantic.min.css';
 
 import RawCalendar from 'components/RawCalendar/RawCalendar';
+import parseDate from 'components/RawCalendar/parseDate';
 
-const parseDate = (dateInfo) => {
-  const dateParser = /^([A-Za-z]+) ([A-Za-z]+) (\d+) (\d+)/;
-  const month = dateInfo.match(dateParser)[2];
-  const date = dateInfo.match(dateParser)[3];
-  const year = dateInfo.match(dateParser)[4];
-
-  const months = {
-    Jan: '01',
-    Feb: '02',
-    Mar: '03',
-    Apr: '04',
-    May: '05',
-    Jun: '06',
-    Jul: '07',
-    Aug: '08',
-    Sep: '09',
-    Oct: '10',
-    Nov: '11',
-    Dec: '12',
-  };
-
-  return `${year}-${months[month]}-${date}`;
-};
 
 class RealCalendar extends Component {
   isDisable = ({ date }) => {
     const { review } = this.props;
-    console.log(date, review);
     return review.filter((r) => (r.date === parseDate(`${date}`))).length === 0;
   }
